third_party/blink/web_tests/virtual/fenced-frame-shadow-dom/wpt_internal/fenced_frame/coop-bcg-swap.https-expected.txt

This is a testharness.js-based test.
FAIL window.name after bcg swap promise_test: Unhandled rejection with value: object "Error: assert_equals: expected "" but got "test""
Harness: the test ran to completion.
